Added 'cutting_tool_edge_angle', 'corner radius', 'flat radius' and 'type' attributes to the cutting tool definition.  I'm trying to get enough attributes that we can draw the cutting tool.  Eventually I want to draw the rotated solid of the cutting tool so that we can subtract that solid from the solids in the model that the operation travels through.  Right now I'm trying to add attributes.
I have based the 'corner radius' and 'flat radius' parameters on my understanding of
what the Adaptive Clearing routines need.  If I'm wrong, the graphics that they
produce will make it clear.

I have also added some code to the properties 'on_set....' routines so that, when
the operator changes one thing, some of the other settings get changed to more
reasonable guesses.  This may prove to be frustrating but I want to avoid having
a bad representation of the cutting tool when I get to the point of using this
description to affect the data model.