Added machine configuration for a basic gantry router that has no spindle control or tool changer.
Fixed problem found when an old .heeks file is openned.  The problem is in the construction of
the various 'parent' ObjList pointers such as m_program->m_tools and m_program->m_speed_references.  It
now checks that these items exist before trying to add to them.  If they don't exist then it
adds them on the fly.

Added double quotes around the python filename so that spaces within the name don't cause
interpretation issues.