Improved the algorithm used to determine corner toolpaths.  It will now behave better for some symetrical shapes.
Fixed problem where the Inlay shapes were being rotated for the first fixture but not for the second fixture (if defined).  It now rotates the various graphics correctly for both the male and/or female fixtures that were defined privately.

For those who are interested...

The Inlay operation behaves better with the rearrangement of toolpaths.  i.e. by machining with the chamfering bit straight down and then cleaning up the edges with the pocketing operation, the results are much neater.

The function of defining two private fixtures also works well.  This allows both the male and female halves to be machined in the same setup with each different colour of wood held in a separate vice.

My next problem is in calculating the areas that are not machined by the clearance tool but could probably still be machined by the chamfering bit.  eg: where the shape (sketch) narrows to a point, the clearance tool used to pocket most of the centre out of the area can't fit down into the narrow area of the shape.  The chamfering bit still works in that area due to the way the corner is formed by raising the chamfering bit (effectively reducing the bit's diameter as it get closer to the tip of the narrow point).  We could use the chamfering bit to machine out the waste material in these narrow pointy shapes but I need to figure out what area is left by the clearance tool.

I have no idea how to do this yet.  If I could access the toolpaths generated by the libArea module then I could work it out but that's all in Python processing which is too late for my C++ decisions to be made.

My other option is to generate some algorithm that figures out how much of the shape's area can be removed with the clearance tool's circular shape.

Personally, I'm waiting for divine intervention.  Failing that, has anyone got any ideas?