Added one extra check before overwriting the m_python_program with the m_textCtrl contents.  The problem is that the wxTextCtrl class (upon which the m_textCtrl is based) can only handle either 30kB or 60kB of information depending on the operating system.  This means that it can sometimes truncate the python program generated.  This is why I added the m_python_program member instead.
If the m_textCtrl is just a truncated version of the m_python_program then the overriting does not occur.  If there are different characters within the m_textCtrl than are in the first 'n' bytes of the m_python_program then the operator must have manually edited the python program.  In this case we do overwrite the m_python_program member.