Removed the warning about using the gradient in the CuttingTool class.  It may be a little early to remove it but it was annoying me.
This is my first run at the ramping entry movement for
a contour operation.  There is still an example where an 'open'
sketch produces an incorrect toolpath within the ramp.  For
'closed' sketches it works fine.  I had left it uncommitted for
a long time trying to find this last error but I decided that it's
no worse than the existing code.  If the user selects a 'plunge'
entry move instead then it works as well as the current version
does.  If they select a 'ramp' entry move then it mostly works
and any errors are quite obvious.

I have also updated the codeblocks project files as this is
what I use for development.